# Cache the key of TimeTicket to prevent instantiation